move tests for space charge into beamfields test
