+++ tests/test_beamfields.py
@@ -1,0 +1,70 @@
+import numpy as np
+
+import pysixtrack
+
+
+def test_track_spacecharge():
+    x_co = 0.1
+    y_co = -0.5
+    sigma_x = 0.5
+    sigma_y = 0.1
+    el1 = pysixtrack.elements.SpaceChargeBunched(
+        number_of_particles=1e11,
+        bunchlength_rms=0.22,
+        sigma_x=sigma_x,
+        sigma_y=sigma_y,
+        length=2.0,
+        x_co=x_co,
+        y_co=y_co,
+    )
+    line_density = el1.number_of_particles / (
+        el1.bunchlength_rms * np.sqrt(2 * np.pi)
+    )
+    el2 = pysixtrack.elements.SpaceChargeCoasting(
+        line_density=line_density,
+        sigma_x=el1.sigma_x,
+        sigma_y=el1.sigma_y,
+        length=el1.length,
+        x_co=el1.x_co,
+        y_co=el1.y_co,
+    )
+
+    # test absolute kick for sigma_x > sigma_y
+    x_offset = 0.2
+    y_offset = -0.5
+    p1 = pysixtrack.Particles()
+    p1.x = x_co + x_offset
+    p1.y = y_co + y_offset
+    p2 = p1.copy()
+    el1.track(p1)
+    el2.track(p2)
+    assert np.isclose(p1.px, 1.33671170365656e-07, atol=1e-15) == True
+    assert np.isclose(p1.py, -6.228154616877465e-07, atol=1e-15) == True
+    assert p1.compare(p2, abs_tol=1e-15)
+
+    # test absolute kick for sigma_y > sigma_x
+    el1.sigma_x = sigma_y
+    el1.sigma_y = sigma_x
+    el2.sigma_x = sigma_y
+    el2.sigma_y = sigma_x
+    p1 = pysixtrack.Particles()
+    p1.x = x_co + y_offset
+    p1.y = y_co + x_offset
+    p2 = p1.copy()
+    el1.track(p1)
+    el2.track(p2)
+    assert np.isclose(p1.px, -6.228154616877465e-07, atol=1e-15) == True
+    assert np.isclose(p1.py, 1.33671170365656e-07, atol=1e-15) == True
+    assert p1.compare(p2, abs_tol=1e-15)
+
+    # test no kick for particle on closed orbit
+    p1 = pysixtrack.Particles()
+    p1.x = el1.x_co
+    p1.y = el1.y_co
+    p2 = p1.copy()
+    el1.track(p1)
+    el2.track(p2)
+    assert np.isclose(p1.px, 0.0, atol=1e-15) == True
+    assert np.isclose(p1.py, 0.0, atol=1e-15) == True
+    assert p1.compare(p2, abs_tol=1e-15)
+
